Add workaround/fix for disc reading
Wii U seems to return no regular file flag for disc partitions, something which was mostly hidden due to a bug in newlib (see https://github.com/devkitPro/newlib/issues/24) and this implementation also assuming that anything non-directory is a regular file.

To fix that it should assume that any file/folder on a disc is a regular file if it's not one of the previous categories.